+++ src/worker/util/changes.basketball.ts
@@ -310,6 +310,11 @@ const changes = [
 		msg:
 			"The Power Rankings page now shows team rankings for individual rating categories.",
 	},
+	{
+		date: "2020-08-01",
+		msg:
+			'Players now have jersey numbers, and teams can retire jersey numbers of former players. <a href="https://basketball-gm.com/blog/2020/08/jersey-numbers/">More details</a>',
+	},
 ];
 
 export default changes;